refactor: change request
