+++ docs/openapi/api-internal-pn-bff-pa-info.yaml
@@ -21,7 +21,7 @@ servers:
   - url: https://webapi.dev.notifichedigitali.it
     description: Ambiente di sviluppo
 tags:
-  - name: InstitutionProducts
+  - name: InfoPa
     description: Invocazioni per sapere i portali disponibili di un ente
 
 paths:
@@ -34,7 +34,7 @@ paths:
 #      security:                                      # ONLY EXTERNAL
 #        - bearerAuth: [ ]                            # ONLY EXTERNAL
       tags:
-        - InstitutionAndProduct
+        - InfoPa
       parameters: # NO EXTERNAL
         - $ref: 'common-refs.yaml#/components/parameters/uidAuthFleet'        # NO EXTERNAL
         - $ref: 'common-refs.yaml#/components/parameters/cxTypeAuthFleet'     # NO EXTERNAL
@@ -69,7 +69,7 @@ paths:
 #      security:                                      # ONLY EXTERNAL
 #        - bearerAuth: [ ]                            # ONLY EXTERNAL
       tags:
-      - InstitutionAndProduct
+      - InfoPa
       parameters: # NO EXTERNAL
         - $ref: 'common-refs.yaml#/components/parameters/uidAuthFleet'        # NO EXTERNAL
         - $ref: 'common-refs.yaml#/components/parameters/cxTypeAuthFleet'     # NO EXTERNAL
@@ -94,3 +94,50 @@ paths:
             application/problem+json:
               schema:
                 $ref: 'common-refs.yaml#/components/schemas/Problem'
+  
+  '/bff/v1/groups':
+    get:
+      summary: Get groups
+      description: |-
+        This operation retrieves the groups from Self Care.
+      operationId: getGroupsV1
+      #      security:                                      # ONLY EXTERNAL
+      #      - bearerAuth: [ ]                            # ONLY EXTERNAL
+      tags:
+      - InfoPa
+      parameters: # NO EXTERNAL
+      - $ref: 'common-refs.yaml#/components/parameters/uidAuthFleet'                          # NO EXTERNAL
+      - $ref: 'common-refs.yaml#/components/parameters/cxIdAuthFleet'                         # NO EXTERNAL
+      - $ref: 'common-refs.yaml#/components/parameters/cxGroupsAuthFleet'                     # NO EXTERNAL
+      - name: statusFilter
+        description: Se valorizzato indica di tornare solo i gruppi nello stato passato
+        in: query
+        required: false
+        schema:
+          $ref: '#/components/schemas/BffPaGroupStatus'
+      responses:
+        '200':
+          description: OK
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/BffPaGroupList'
+        '400':
+          description: Invalid input
+          content:
+            application/problem+json:
+              schema:
+                $ref: 'common-refs.yaml#/components/schemas/Problem'
+        '500':
+          description: Internal Server Error
+          content:
+            application/problem+json:
+              schema:
+                $ref: 'common-refs.yaml#/components/schemas/Problem'
+
+components:
+  schemas:
+    BffPaGroupStatus:
+      $ref: 'https://raw.githubusercontent.com/pagopa/pn-external-registries/f49e1a815e81a70834c21c95c0fb7e9e8d39c054/docs/openapi/pn-selfcare-internal-v1.yaml#/components/schemas/PaGroupStatus'
+    BffPaGroupList:
+      $ref: 'https://raw.githubusercontent.com/pagopa/pn-external-registries/f49e1a815e81a70834c21c95c0fb7e9e8d39c054/docs/openapi/pn-selfcare-internal-v1.yaml#/components/schemas/PaGroupList'
